gsl: Skip debug token for Register Events opcode
GPR packet token field is used by APM modules when raising events.
GSL update for debug tokens created an issue, as the modules need to
to match the same tokens to events. GSL previously was always setting
token to 0 in for events. This patch removes tokens specifically for the
opcode which causes the issue.

Signed-off-by: Aditya Rathi <quic_aditrath@quicinc.com>